fit : разбиение логики экрана
